+++ src/main/java/com/github/dakusui/scriptiveunit/drivers/Arith.java
@@ -1,13 +1,12 @@
 package com.github.dakusui.scriptiveunit.drivers;
 
 import com.github.dakusui.scriptiveunit.annotations.Scriptable;
+import com.github.dakusui.scriptiveunit.model.form.Form;
 import com.github.dakusui.scriptiveunit.model.form.FormList;
 import com.github.dakusui.scriptiveunit.model.session.Stage;
-import com.github.dakusui.scriptiveunit.model.form.Form;
 import com.github.dakusui.scriptiveunit.utils.CoreUtils;
 
 import java.math.BigDecimal;
-import java.util.List;
 import java.util.function.Function;
 
 import static java.math.MathContext.DECIMAL128;
@@ -48,7 +47,7 @@ public class Arith {
     BigDecimal ret = null;
     for (Object eachNumber : numbers) {
       BigDecimal each = CoreUtils.toBigDecimal((Number)
-          ((Form)eachNumber).apply(stage));
+          ((Form) eachNumber).apply(stage));
       if (ret == null) {
         ret = requireNonNull(CoreUtils.toBigDecimal(each));
       } else {
