--- src/main/java/com/github/dakusui/scriptiveunit/utils/CliUtils.java
@@ -1,5 +1,0 @@
-package com.github.dakusui.scriptiveunit.utils;
-
-public enum CliUtils {
-  ;
-}
